+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added meta data (formulas, ingredients,...) count for users in the admin dashboard
 - Monitor user changes in users admin dashboard
 - Added last login for users
+- Formula making improvements
 
 ### Version 12.5
 - Code clean-up