+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v12.6
 --------------------------
+- Formula making improvements
 - Added meta data (formulas, ingredients,...) count for users in the admin dashboard
 - This release may include more changes, for full details please refer to the CHANGELOG